Rename oso.py in flask oso.
