+++ languages/python/flask_oso/flask_oso/__init__.py
@@ -1,6 +1,6 @@
 __version__ = '0.0.0'
 
-from .oso import FlaskOso
+from .flask_oso import FlaskOso
 from .decorators import authorize, skip_authorization
 
 __all__ = ['FlaskOso', 'authorize', 'skip_authorization']
